rename classes and props
