+++ src/main/java/ee/openeid/siva/demo/configuration/SivaServiceProperties.java
@@ -23,9 +23,9 @@ import org.springframework.boot.context.properties.ConfigurationProperties;
 @ConfigurationProperties(prefix = "siva.service")
 public class SivaServiceProperties {
     private static final String DEFAULT_SERVICE_URL = "http://localhost:8080";
-    private String jsonServicePath = "/validate";
-    private String jsonHashcodeServicePath = "/validateHashcode";
-    private String jsonDataFilesServicePath = "/getDataFiles";
+    private String servicePath = "/validate";
+    private String hashcodeServicePath = "/validateHashcode";
+    private String dataFilesServicePath = "/getDataFiles";
     private String serviceHost = DEFAULT_SERVICE_URL;
     private String trustStore = "siva_server_truststore.p12";
     @SuppressWarnings("squid:S2068") //default password
