+++ src/main/java/ee/openeid/siva/demo/siva/SivaDataFilesServiceClient.java
@@ -19,7 +19,6 @@ package ee.openeid.siva.demo.siva;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import ee.openeid.siva.demo.cache.UploadedFile;
 import ee.openeid.siva.demo.configuration.SivaServiceProperties;
-import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.ResourceAccessException;
@@ -28,7 +27,8 @@ import org.springframework.web.client.RestTemplate;
 import java.io.IOException;
 
 @Service
-public class SivaJSONValidationServiceClient implements ValidationService {
+public class SivaDataFilesServiceClient implements DataFilesService {
+
     private static final int GENERIC_ERROR_CODE = 101;
 
     private SivaServiceProperties properties;
@@ -36,26 +36,19 @@ public class SivaJSONValidationServiceClient implements ValidationService {
     private SivaValidationServiceErrorHandler errorHandler;
 
     @Override
-    public String validateDocument(final String policy, final String report, final UploadedFile file) throws IOException {
+    public String getDataFiles(UploadedFile file) throws IOException {
         if (file == null) {
             throw new IOException("Invalid file object given");
         }
 
         final String base64EncodedFile = file.getEncodedFile();
 
-        final ValidationRequest validationRequest = new ValidationRequest();
-        validationRequest.setDocument(base64EncodedFile);
-        if (StringUtils.isNotBlank(policy))
-            validationRequest.setSignaturePolicy(policy);
-        if (StringUtils.isNotBlank(report))
-            validationRequest.setReportType(report);
-        final String filename = file.getFilename();
-        validationRequest.setFilename(filename);
-
+        final DataFilesRequest dataFilesRequest = new DataFilesRequest();
+        dataFilesRequest.setDocument(base64EncodedFile);
+        dataFilesRequest.setFilename(file.getFilename());
         try {
             restTemplate.setErrorHandler(errorHandler);
-            String fullUrl = properties.getServiceHost() + properties.getJsonServicePath();
-            return restTemplate.postForObject(fullUrl, validationRequest, String.class);
+            return restTemplate.postForObject(properties.getServiceHost() + properties.getDataFilesServicePath(), dataFilesRequest, String.class);
         } catch (ResourceAccessException ce) {
             String errorMessage = "Connection to web service failed. Make sure You have configured SiVa web service correctly";
             return new ObjectMapper().writer().writeValueAsString(new ServiceError(GENERIC_ERROR_CODE, errorMessage));
